requirements.txt
amqp==2.5.2
asgiref==3.2.3
beautifulsoup4==4.8.2
billiard==3.6.1.0
certifi==2019.11.28
chardet==3.0.4
Django==3.0.2
idna==2.9
kombu==4.6.7
lxml==4.5.0
micawber==0.5.1
pyembed==1.3.3
pytube3==9.6.4
pytz==2019.3
requests==2.23.0
soupsieve==1.9.5
sqlparse==0.3.0
typing-extensions==3.7.4.1
urllib3==1.25.8
vine==1.3.0
youtube-dl==2020.2.16
